Moved legal components into a subdirectory.
