+++ assets/components/legal/contribLegal/contribLegal.jsx
