+++ assets/components/legal/contribLegal/contribLegal.scss
