+++ assets/components/legal/termsPrivacy/termsPrivacy.jsx
