+++ assets/components/legal/termsPrivacy/termsPrivacy.scss
