+++ assets/stylesheets/main.scss
@@ -12,7 +12,7 @@
 
 @import '../components/bodyCopy/bodyCopy';
 @import '../components/contribAmounts/contribAmounts';
-@import '../components/contribLegal/contribLegal';
+@import '../components/legal/contribLegal/contribLegal';
 @import '../components/ctaCircle/ctaCircle';
 @import '../components/ctaLink/ctaLink';
 @import '../components/errorMessage/errorMessage';
@@ -28,7 +28,7 @@
 @import '../components/radioToggle/radioToggle';
 @import '../components/payPalExpressButton/payPalExpressButton';
 @import '../components/stripePopUpButton/stripePopUpButton';
-@import '../components/termsPrivacy/termsPrivacy';
+@import '../components/legal/termsPrivacy/termsPrivacy';
 @import '../components/textInput/textInput';
 @import '../components/testUserBanner/testUserBanner';
 @import '../components/secure/secure';
